refactor(ClientFactory): refactor OAuth2 handler
